Roles y permisos de ususarios

+++ app/Http/Controllers/UserController.php
@@ -1,13 +1,18 @@
 <?php
 
 namespace App\Http\Controllers;
-use App\Models\User;
+
+
 use App\Http\Requests\UserRequest;
 use Illuminate\Http\Request;
+use App\Models\User;
 use Carbon\Carbon;
 use PDF;
 use App\Exports\UserExport;
 use Maatwebsite\Excel\Facades\Excel;
+
+use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
@@ -23,8 +28,9 @@ class UserController extends Controller
 
     public function index()
     {
-        $datos['Users']=User::paginate(10);
-        return view('Usuarios.usuariosindex',$datos);
+        abort_if(Gate::denies('user_index'), 403);
+        $users=User::paginate(10);
+        return view('Usuarios.usuariosindex', compact('users'));
     }
 
     public function pdf()
@@ -33,8 +39,8 @@ class UserController extends Controller
         $hoy=$mytime->toDateTimeString();
         $direccion="Colonia Humuya, Avenida Altiplano, Calle Poseidón, 11101";
 
-        $User = User::paginate();
-        $pdf = PDF::loadView('usuarios.userpdf',compact('User','hoy'));
+        $user = User::paginate();
+        $pdf = PDF::loadView('usuarios.userpdf',compact('user','hoy'));
         //$pdf->loadHTML ('<h1>Test</h1>');
 
         //return $pdf->stream();
@@ -48,6 +54,7 @@ class UserController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('user_create'), 403);
         return view('Usuarios.create');
     }
 
@@ -61,8 +68,8 @@ class UserController extends Controller
     {
         //
         //$datosUser = request()->all();
-        $User = request()->except('_token');
-        user::insert($User);
+        $user = request()->except('_token');
+        User::insert($user);
         alert()->success('Usuario guardado correctamente');
         
         return redirect()->route('usuarios.index');
@@ -75,9 +82,12 @@ class UserController extends Controller
      * @param  \App\Models\User  $user
      * @return \Illuminate\Http\Response
      */
-    public function show(User $user)
+    public function show($id)
     {
-        //
+        abort_if(Gate::denies('user_show'), 403);
+        $user = User::findOrFail($id);
+        $user->load('roles');
+        return view('usuarios.show', compact('user'));   
     }
 
     /**
@@ -86,10 +96,14 @@ class UserController extends Controller
      * @param  \App\Models\User $user
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
+    public function edit($id)    
     {
-        /*$User=User::findOrFail($User);
-        return redirect()->route('usuarios.index');*/
+        abort_if(Gate::denies('user_edit'), 403);
+        $user=User::findOrFail($id);
+        //abort_if(Gate::denies('user_edit'), 403);
+        $roles = Role::all()->pluck('name', 'id');
+        $user->load('roles');
+        return view('usuarios.edit', compact('user', 'roles'));
     }
 
     /**
@@ -99,24 +113,19 @@ class UserController extends Controller
      * @param  \App\Models\User $user
      * @return \Illuminate\Http\Response
      */
-    public function update(UserRequest $request, $id)
+    public function update(Request $request, $id)
     {
-        $User= request()->except(['_token','_method']);
-        /*User::where('id','=',$User)->update($User);*/
-        User::where('id','=',$id)->update($User);
-       /*$User = request()->except('_token');
-        User::update($User);
-        /*$User=User::findOrFail($User);*/
-        alert()->success('Usuario Actualizado correctamente');
-        return redirect()->route('usuarios.index');
+        //$User= request()->except(['_token','_method']);
 
-        
-        /*$datosUser = request()->except(['_token','_method']);
-        User::where('id','=',$id)->update($datosUser);
-        alert()->success('Producto guardado correctamente');
-        
-        $User=User::findOrFail($id);
-        return view('usuarios');*/
+        //User::where('id','=',$id)->update($User);
+
+        $user = User::findOrFail($id);
+        $data = $request->only('name', 'username', 'email');
+
+        $roles = $request->input('roles', []);
+        $user->syncRoles($roles);
+
+        return redirect()->route('usuarios.edit', $user)->with('info','Se asigno los roles correctamente');
 
         
     }
@@ -129,8 +138,19 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        abort_if(Gate::denies('user_destroy'), 403);
+        $user = User::findOrFail($id);
+        if (auth()->user()->id == $user->id){
+            return redirect()->route('usuarios.usuariosindex');
+        }
+        
         User::destroy($id);
+        
+        
+        
         return redirect('usuarios')->with('mensaje','Usuario Borrado');
+
+
 
     }
 }
